+++ src/server/js/index.js
@@ -3,7 +3,6 @@
  */
 import '../../common/js/index.js';
 
-import './accounts.js';
 import './check_npms.js';
 import './functions.js';
 import './maintain.js';