+++ ChangeLog.md
@@ -14,6 +14,7 @@
     - Add missing pwix:ui-utils and ui-fontawesome6 dependencies
     - Add scopeLabelFn, scopesFn and scopesPub configuration parameters to get known scopes
     - Define Roles.EditPanel.global(), .scoped() and .roles() function to get back the edition results
+    - Define Roles.setUserRoles() new function
 
 ### 1.4.0